DONE LOGIC LoaiQuyTrinh

+++ app/Http/Controllers/LoaiQuyTrinhController.php
@@ -10,6 +10,7 @@ use App\Models\NguyenVatLieu;
 use App\Models\NguyenVatLieuLoaiQuyTrinh;
 use Illuminate\Http\Request;
 
+
 class LoaiQuyTrinhController extends Controller
 {
     /**
@@ -19,22 +20,11 @@ class LoaiQuyTrinhController extends Controller
      */
     public function index(Request $request, LoaiQuyTrinh $quytrinh)
     {
-        
-
         $search = $request->get('q');
         $quytrinh = $quytrinh->query()
             ->where('MaLoaiQuyTrinh', 'like', '%' . $search . '%')
             ->paginate(10);
 
-        /* foreach($quytrinh as $quytrinh){
-            echo $quytrinh->MaLoaiQuyTrinh;
-            echo '<br>';
-            foreach($quytrinh->nguyenVatLieu as $nguyenvatlieu){
-                echo $nguyenvatlieu->TenNguyenVatLieu.',';
-            }
-            echo '<hr>';
-        }
-        dd(); */
         return view('Qlqt.index', [
             'quytrinh' => $quytrinh,
             'search'  => $search,
@@ -48,11 +38,16 @@ class LoaiQuyTrinhController extends Controller
      */
     public function create(Request $request, LoaiQuyTrinh $quytrinh)
     {
-        $nguyenvatlieu = NguyenVatLieu::get();
+        // $nguyenvatlieu = NguyenVatLieu::get();
         $chatlieu = ChatLieu::get();
+        $nguyenvatlieudatchuan = NguyenVatLieu::join('chitietnhapkho', 'chitietnhapkho.MaNguyenVatLieu', '=', 'nguyenvatlieu.MaNguyenVatLieu')
+            ->join('nhapkho', 'nhapkho.MaNhapKho', '=', 'chitietnhapkho.MaNhapKho')
+            ->join('tinh_trang_nhap_kho', 'tinh_trang_nhap_kho.MaNhapKho', '=', 'nhapkho.MaNhapKho')
+            ->where('tinh_trang_nhap_kho.TinhTrang', '=', 1)->get();
+
         return view('Qlqt.create', [
-            'nguyenvatlieu' => $nguyenvatlieu,
-            'chatlieu' => $chatlieu,
+            'nguyenvatlieu' => $nguyenvatlieudatchuan,
+            'chatlieu' => $chatlieu
         ]);
     }
 
@@ -64,17 +59,14 @@ class LoaiQuyTrinhController extends Controller
      */
     public function store(StoreLoaiQuyTrinhRequest $request, LoaiQuyTrinh $quytrinh)
     {
-        // $quytrinh->create($request->validated());
-        // return redirect()->route('quytrinh.index')->with('success', 'Đã thêm thành công');
-
         $quytrinh->fill($request->all()); // Lấy hết dữ liệu
-        $quytrinh->fill($request->except('_token')); // Lấy hết dữ liệu ngoại trừ thuộc tính _token
+        $quytrinh->fill($request->except('_token', 'NguyenVatLieu')); // Lấy hết dữ liệu ngoại trừ thuộc tính _token
         $quytrinh->save();
 
-        foreach ($request->NguyenVatLieu as $value) {
+        foreach ($request->NguyenVatLieu as $nguyenVatLieu) {
             $nguyenVatLieuLoaiQuyTrinh = new NguyenVatLieuLoaiQuyTrinh();
             $nguyenVatLieuLoaiQuyTrinh->setLoaiQuyTrinh($request->MaLoaiQuyTrinh);
-            $nguyenVatLieuLoaiQuyTrinh->setNguyenVatLieu($value);
+            $nguyenVatLieuLoaiQuyTrinh->setNguyenVatLieu($nguyenVatLieu);
             $nguyenVatLieuLoaiQuyTrinh->save();
         }
 
@@ -100,25 +92,42 @@ class LoaiQuyTrinhController extends Controller
      */
     public function edit(LoaiQuyTrinh $quytrinh)
     {
-        $nguyenvatlieu = NguyenVatLieu::get();
+        // $nguyenvatlieu = NguyenVatLieu::get();
+        $chatlieu = ChatLieu::get();
+        $nguyenvatlieudatchuan = NguyenVatLieu::join('chitietnhapkho', 'chitietnhapkho.MaNguyenVatLieu', '=', 'nguyenvatlieu.MaNguyenVatLieu')
+            ->join('nhapkho', 'nhapkho.MaNhapKho', '=', 'chitietnhapkho.MaNhapKho')
+            ->join('tinh_trang_nhap_kho', 'tinh_trang_nhap_kho.MaNhapKho', '=', 'nhapkho.MaNhapKho')
+            ->where('tinh_trang_nhap_kho.TinhTrang', '=', 1)->get();
+
         return view('Qlqt.edit', [
             'data' => $quytrinh,
-            'nguyenvatlieu' => $nguyenvatlieu,
+            'chatlieu' => $chatlieu,
+            'nguyenvatlieu' => $nguyenvatlieudatchuan,
         ]);
     }
 
     public function update(UpdateLoaiQuyTrinhRequest $request, LoaiQuyTrinh $quytrinh)
     {
         // dd($request);
-        $quytrinh->update($request->validated());
-        $deleteValueCu = NguyenVatLieuLoaiQuyTrinh::where('MaLoaiQuyTrinh', $request->MaLoaiQuyTrinh)->delete();
-        foreach ($request->NguyenVatLieu as $value) {
-            $nguyenVatLieuLoaiQuyTrinh = new NguyenVatLieuLoaiQuyTrinh();
-            $nguyenVatLieuLoaiQuyTrinh->setLoaiQuyTrinh($request->MaLoaiQuyTrinh);
-            $nguyenVatLieuLoaiQuyTrinh->setNguyenVatLieu($value);
-            $nguyenVatLieuLoaiQuyTrinh->save();
+        if ($request->NguyenVatLieu) {
+            $quytrinh->update($request->validated());
+            $deleteValueCu = NguyenVatLieuLoaiQuyTrinh::where('MaLoaiQuyTrinh', $request->MaLoaiQuyTrinh)->delete();
+            foreach ($request->NguyenVatLieu as $value) {
+                $nguyenVatLieuLoaiQuyTrinh = new NguyenVatLieuLoaiQuyTrinh();
+                $nguyenVatLieuLoaiQuyTrinh->setLoaiQuyTrinh($request->MaLoaiQuyTrinh);
+                $nguyenVatLieuLoaiQuyTrinh->setNguyenVatLieu($value);
+                $nguyenVatLieuLoaiQuyTrinh->save();
+            }
+            
+            return redirect()->route('quytrinh.index')->with('edited', 'Sửa thành công');
+        }else {
+            $nguyenVatLieuLoaiQuyTrinh = $request->validate([
+                'MaNguyenVatLieu' => 'required',
+            ],
+            [
+                'MaNguyenVatLieu.required' => 'Nguyên vật liệu phải chọn nha',     
+            ]);
         }
-        return redirect()->route('quytrinh.index')->with('edited', 'Sửa thành công');
     }
 
     /**
